Fix: FBX fails to export materials on geometry nodes objects
For only the FBX exporter, fixes #104516.

The exporter would always get the materials to export from the original,
non-evaluated, object, but Geometry nodes, when applied, replace the
materials of the object and mesh/curve/etc., so the wrong materials
would be exported.

Now, when modifiers are applied on export, the materials of the
evaluated object are checked against the materials of the original
object. When the materials differ, the materials of the evaluated object
are set into the `override_materials` of the ObjectWrapper, causing its
`materials` property (refactored from the `material_slots` property) to
return them.

Pull Request: https://projects.blender.org/blender/blender-addons/pulls/104530