FBX IO: Vertex normals access with new `vertex_normals` property
Blender 3.1 moved vertex normals to a separate array. The old API still
works for now, but is slower and may be removed in 4.0, so this patch
updates FBX IO to use the new `vertex_normals` property.

This patch makes no changes to the import or export of FBX files.

Pull Request: https://projects.blender.org/blender/blender-addons/pulls/104651