Sun Position: replace DMS label by the Enter Coordinates field
The Latitude and Longitude could be displayed in a label underneath
the interface sliders in a Degrees, Minutes, Seconds form.

This was redundant with the field above used to input coordinates to
be parsed, as it can also display the coords in a canonical form. The
coordinates can either be parsed from a string, or from the sliders
themselves.