Sun Position: remove show_daylight_savings preference
This option, like others previously removed, may lead to faulty
behavior. Since the scene depends on user preferences, a given file
may not give the same result on one installation as on another. One
possible solution would be take the value into account even if the
checkbox is hidden, but it could lead to a confusing one-hour
discrepancy in the UI.

The best is to remove the option altogether, so that the checkbox is
always shown. It does not take too much space.

Also bump minimum Blender version to 3.2 as the gpu API has changed
since 3.1 and the add-on is no longer compatible with it.