FBX IO: Vertex position access with attributes
Blender 3.5 moved vertex positions to a generic attribute. The old API
still works for now, but is slower and may be removed in 4.0, so this
patch updates FBX IO to use the new "position" attribute.

This patch makes no changes to the import or export of FBX files.

Pull Request: https://projects.blender.org/blender/blender-addons/pulls/104647