FBX IO: Corner vert/edge and edge verts access with attributes
Blender 3.6 moved corner (loop) vertex index and edge index, and edge
vertices to generic attributes. The old API still works for now, but is
slower and may be removed in 4.0, so this patch updates FBX IO to use
the new ".corner_vert", ".corner_edge" and ".edge_verts" attributes.

numpy.astype() with copy=False does not create a new view in the desired
dtype when the original dtype is the same kind, itemsize and byteorder,
but has a different character code, e.g. 'i' and 'l'. This is
problematic because Blender can be picky about the character code. The
astype_view_signedness() utility function has been updated to create
such views instead of falling back to numpy.astype().

This patch makes no changes to the import or export of FBX files.

Pull Request: https://projects.blender.org/blender/blender-addons/pulls/104648