VDM brush baker
An add-on that simplifies the creation of vector displacement brushes in Blender.

The add-on panel can be found in the tool panel.
'Create sculpting plane' will create a grid with two multires subdivisions applied. It can be used to sculpt the vector displacement.
'Render and create brush' will as the name implies create VDM from the sculpting plane plus a sculpting brush that uses the VDM.
The material that is needed for baking VDMs is created via code.

Pull Request: https://projects.blender.org/blender/blender-addons/pulls/104580