Fix #104576: FBX export of Edit Mode Objects while Blender is not in Edit Mode can fail.
All Objects to be exported are now checked for not being in Edit Mode
prior to exporting.

In the rare case that an Object is found to be in Edit Mode, it is
temporarily added to the current view layer, temporarily set as active
and then set to Object Mode.

If an Object cannot be set to Object mode for some reason, an error is
reported to the Operator and exporting is aborted.

Pull Request: https://projects.blender.org/blender/blender-addons/pulls/104633