Fix #104672: FBX Byte type unsupported
The Byte type "Z" was missing from the supported FBX types, which
would result in an error attempting to parse an FBX containing Byte
properties during import or conversion with fbx2json.

The new variables for the Byte type refer to it as "int8", to reduce the
chance of mixing it up with the existing "bytes" attributes and
functions.

The importer and exporter neither read nor write any Byte properties, so
no changes are made to import or export.

Pull Request: https://projects.blender.org/blender/blender-addons/pulls/104681