Fix disparity of FBX custom properties import and export
Light, camera, bone and pose bone properties are now imported to match
the exporter behaviour.

Image custom properties are now exported to match the importer
behaviour, except for Image custom properties that are rarely imported
from a Texture Node instead of a Video Node, which are always exported
to a Video Node.

This resolves one of the issues in #104677.

Pull Request: https://projects.blender.org/blender/blender-addons/pulls/104702