Fix #104707: FBX edge creases fail to be imported/exported
The MeshEdge.crease API has been removed in 4.0, replaced with the
"crease_edge" attribute and the Python defined
Mesh.edge_creases/edge_creases_ensure()/edge_creases_remove().

This patch replaces use of the old MeshEdge.crease API with accessing
the "crease_edge" attribute with the Python defined property and
function.

The intention of this patch is not to perform any functional changes
compared to when the older MeshEdge.crease API was used, so while it may
be possible to skip exporting creases when there aren't any, this has
been left as a "todo" comment for now.

Pull Request: https://projects.blender.org/blender/blender-addons/pulls/104709