FBX Import: Read known sequential bytes simultaneously
The start of each element and the start of each array property have a
consistent series of values with fixed sizes. Reading these series of
values with a single read call and single unpack call is slightly faster
than reading them individually.

This reduces the time taken to parse fbx files to about 90% to 95% of
the original time. Array decompression makes up about half of the time
taken currently, but this could be multithreaded, in which case, this
patch would reduce the time to about 80% to 90% instead.

Pull Request: https://projects.blender.org/blender/blender-addons/pulls/104823