io_scene_3ds: Added spotlight aspect and projector
Added aspect and projector chunks to importer and exporter.
Now it is possible to export an image from the light tree if connected to an emission shader.
Import will add an image node with the gobo / projector image and connects it to emission shader.
The scale of the x and y axes of a spot can now be imported and exported in a aspect ratio chunk.

Pull Request: https://projects.blender.org/blender/blender-addons/pulls/104813