Pose Library: Update to use the asset shelf (when enabled)
A dedicated region to display poses was already on the wish-list during initial
discussions for the new asset based pose library design. To minimize
development time, it was decided to use the sidebar for the start instead.
Problem is that this space is used for bone pickers, selection set access and
all kinds of other features, typically from add-ons. So this is precious space
for all kinds of important functionality, and animators end up scrolling and
switching tabs a lot, slowing down the workflow.

The recently introduced asset shelf addresses this, see 98142f5e35.  As of
writing, the asset shelf is still an experimental feature. It will probably be
made non-experimental soon. The updates to the pose library add-on use the old
UI in the sidebar if the asset shelf is disabled.

Pull Request: https://projects.blender.org/blender/blender-addons/pulls/104546