Storm Hydra render addon
Storm is a realtime renderer part of USD. When enabling the add-on,
it can be chosen as an alternative to Cycles, Eevee or Workbench.

Compared to Eevee its capabilities are limited, as it was designed
mostly for previews. It serves two main purposes:
* In production pipelines where Blender outputs USD files, it provides
  a preview of how the scene will be exported to other applications.
* As a reference for Blender, add-on and studio pipeline developers
  to validate USD and Hydra integration.

Co-authored-by: Georgiy Markelov <dagerd@noreply.localhost>
Co-authored-by: Vasyl-Pidhirskyi <vpidhirskyi@gmail.com>
Co-authored-by: Brecht Van Lommel <brecht@blender.org>

Pull Request: https://projects.blender.org/blender/blender-addons/pulls/104597