Copy Render Settings: Add support for Cycles, Eevee and Workbench settings.
In addition to `Scene.render` properties, add properties from
`Scene.cycles`, `Scene.eevee` and `Scene.display.shading` to the list of
copyable data.

Note that the list of propoerties is now fairly gigantic, the search
field at the bottom of the list becomes a critical help.

Better ways to control what is shown and to filter properties should be
added at some point, when there is time for such development.

Fix #104771.