Pose Library: Always use asset shelf, remove old UI (hint at new one)
The asset shelf is not an experimental feature anymore since 7bb693581f,
so use this for this for the pose library UI. The old sidebar UI (using
the asset-view template) is removed.

Since this is the only use-case of the asset shelf for the start,
discoverability is a bit low still. So this adds some info text to where
the old asset-view template was, that hints at the asset shelf. There's
also a button to toggle the asset shelf. This design needs to be
evaluated with the animation module still, some other way to make it
more discoverable may be prefered for the actual release.