Rigify: add a warning if feature set requires newer Blender version
Users often install a Rigify feature set that was created for a newer
version of Blender. This is a source of frustration and bug reports.

This commit adds a basic warning to the feature set panel if the
active feature set was meant to be used with a newer Blender version.

Alleviates #97622.

Pull Request: https://projects.blender.org/blender/blender-addons/pulls/104631