FBX IO: Speed up animation import using NumPy
This patch updates much of the animation import code to replace
iteration with NumPy vectorized functions and can result in a
significant speedup for files that are mostly animation.

An FBX file containing an armature with 65 bones and a single animation
with 1523 frames (each one keyframed) animating 47 bones (144 animated
channels in the FBX file) goes from a 1.96s import to a 1.01s import for
me with this patch.

The majority of the remaining animation import time after this patch is
spent converting FBX transformation animations to Blender transformation
animations. Much of this can also be sped up with NumPy, but that will
be submitted separately once it is finished. For this reason, the new
`_transformation_curves_gen` function is only intended as an interim to
facilitate this patch until the transformation animation conversion is
also updated to use NumPy.

# Changes to the import of valid animations:
Previously, all imported animation curves for a single Action would
interpolate the keyframes of every other animation curve imported for
that action. This is needed for imported Object/PoseBone transformation
animations, but is unnecessary for other imported animations.

This was unlikely to be noticed because it only affected material color
animations and animations that affected both the `lens` and
`focus_distance` properties of a camera.

Now with this patch, only the transformation animation curves
interpolate the keyframes of every other transform animation curve.

This change is a requirement to be able to support importing other
animation curves in the future, such as those for custom properties.

# Changes to the import of questionably valid animations:
FBX's default animation system only uses the first animation curve
assigned to a property's channel, but the specification does allow for
more than one curve per channel.

Previously, Blender would combine these curves together, but in such a
way that each additional imported curve would replace all the values of
prior imported curves at times less than the maximum keyframe time of
the additional imported curve (interpolating values for the keyframes
that don't exist in the additional imported curve).

Now when there are multiple animation curves assigned to a channel, the
additional imported curves only replace values at their keyframe times,
no interpolation for other keyframe times takes place.

# Changes to the import of invalid animations:
An animation curve is invalid if its keyframe times are not strictly
increasing. Some imported invalid curves result in extrapolating values.
These extrapolated values are still discarded with this patch, but
previously they would be replaced with a default value. With this patch,
the extrapolated values and their keyframe times are discarded entirely.

Pull Request: https://projects.blender.org/blender/blender-addons/pulls/104856