Rigify: add an api to use runtime-evaluated expressions in panel scripts.
Allow including expressions that depend on the current values of bone
properties into the generated UI script for the panel. This can be used
for disabling buttons based on custom conditions, but the expressions
can be used anywhere arbitrary values are allowed.

In order to make the interface as clean as possible, this is implemented
via a class that wraps a string that is returned as its repr(), and
implements operators and methods for augmenting that expression string.