FBX IO: Speed up cluster export: Remove unused sorting
Sorting the weights and vertex group indices on each vertex was added in
a47e870349 as part of an option to limit the maximum number of exported
weights per vertex. This feature was later removed, but the extra code
that added the sorting was not.

This patch removes the sorting because it is not needed and is quite
expensive. Setting all the weights and indices into `vgroups` now runs
in about 40% of the original time.

The overall speedup from this patch depends very much on what data other
than the clusters needs exporting, but I get a reduction in total export
time between 7.5% and 15% for non-animated humanoid character models.

Pull Request: https://projects.blender.org/blender/blender-addons/pulls/104922