FBX IO: Speed up animation simplification using NumPy
This patch re-writes animation curve simplification on export to use
NumPy where possible for performance.

An export of a 1523 frame animation of a 65 bone armature animating 47
of the bones (almost entirely rotation animation) goes from a 2.9s
export to a 2.2s export for me with this patch (using the default
Simplify value of 1.0).

There are no expected changes to the contents of exported files with
this patch.

Pull Request: https://projects.blender.org/blender/blender-addons/pulls/104904