UV Export: I18n: disambiguate "Modified" message
The "Modified" msgid is used in both
`bpy.types.UV_OT_export_layout.modified` and the Splash screen
tooltip. In the former, it's used for an option of transformed mesh UV
by modifier, but the latter is for the file update time.

This commit adds the Mesh context to the UV operator property.

Reported by Satoshi Yamasaki in blender/blender#43295.