Node Wrangler: Remove Switch Type operator
Remove Switch Type operator as it is no longer operational
in all node editors, and "the fix" may be a native implementation
anyway (see #101514).

When the shortcut `Shift + S` is pressed, the notice of removal with
the future plan of native implementation is displayed. The native
implementation might use the shortcut.

Pull Request: https://projects.blender.org/blender/blender-addons/pulls/104957