Fix #104991: Node Wrangler Merge Nodes shadows other keymaps (Ctrl+,)
Node Wrangler's Merge Nodes operator uses many shortcuts to merge
inputs in different ways. One of those shortcuts is Ctrl + , (comma)
which was not used by default until Blender 4.0. In this version it
was affected the Open Preferences Operator, resulting in a conflict.

In many cases, the operator will have no action, such as when no node
tree is active, or when no node is selected. In those cases, the
operator can simply return {'PASS_THROUGH'} and let the shortcut be
handled by another operator.

A cleaner solution to this issue would be to refactor the operator so
that the list of nodes to act on is gathered in the poll() method, and
not call the operator at all if there is nothing to do, but this
refactor would require deeper changes.