Fix #105009: AnimAll: Error when inserting key on string attribute
String attributes cannot be animated, but the generator function that
gets the paths from the attribute did not handle them properly.
Instead of returning early when passed a string attribute, it would
yield an empty tuple. On the next execution of the generator, some
variables were not initialized which caused an error.