Rigify: allow subclasses of spline_tentacle to lock the Tip Control option.
Since this option effectively represents three major variations
of the control scheme, a sub-class that intends to tune the rig
for a specific use case may benefit from an ability to lock the
option to a specific constant value that fits it best.

This scheme is already used for `BaseSkinChainRig.chain_priority`.