FBX IO: Fix import of shape key animations without any keyframes
Attempting to import a shape key animation without any animation curves
or with only empty animation curves would result in attempting to get
the min/max of an empty list or array.

The animation curve values array is now checked for being non-empty
before attempting to get its minimum and maximum values.

The list of animated shape key deform values is now only created when
there are values to add to it, thus ensuring it is never empty.