FBX IO: Update for new meshes being smooth by default
Blender 3.6 changed newly created meshes to be smooth by default,
therefore, it is no longer necessary to remove the "sharp_face"
attribute if face smoothing was not set when importing a mesh.

The only time when "sharp_face" needs to be removed is if we create the
attribute when attempting to set face smoothing, but setting face
smoothing fails.

No changes to the import of .fbx files are expected with this patch.

Pull Request: https://projects.blender.org/blender/blender-addons/pulls/105039