Fix #105045: Error importing FBX custom properties on placeholder PoseBones
Caused by not understanding an uncommon import case in 716702b97e where
a bone can be created without an FBX Node to import custom properties
from.

A bone in Blender is always created from an FBX Node, so it seemed like
all bones should have an FBX Node that custom properties can be imported
from. However, FBX allows bone and non-bone Nodes to be parented to one
another, which Blender cannot represent exactly due to Armatures only
containing bones. In such cases, the non-bone Node is imported as an
Object, but a bone with the same name is also created (which the Object
is parented to). The newly created bone doesn't have a Node to import
custom properties from because the Node belongs to the imported Object
instead.