Fix #105067: Node Wrangler: cannot preview multi-user material group
When the user tried to use the Node Wrangler operator to preview a
node inside a node group, there was an error. It was introduced in my
refactor commit 1c308ec742 and is fixed by restoring the enumerate()
around the node group's output_sockets.

This fix led to the discovery of another issue: multiple materials can
use a single node group, and in that case multiple `tmp_viewer` are
created so they do not share the same output, but they could not be
distinguished because they were looked up using their `name`s instead
of their unique `identifier`s.