Fix #100973: Node Wrangler: Previewing node if hierarchy not active
Node Wrangler's NWPReviewNode operator used a custom function to get
the currently edited tree, by going through the hierarchy of node
groups starting from the base node tree. This relied on each group
node being active, but it is not always the case if one was deselected
in another editor.

The query for the current node tree now simply uses
`SpaceNodeEditor.edit_tree`.

Note: an unused part of the removed function was getting the path in
the node group hierarchy from the base node tree to the currently
edited one. This can be obtained with `SpaceNodeEditor.path`.